% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGrInformation.R
\name{addGrInformation}
\alias{addGrInformation}
\title{add assembly information to the genomic range}
\usage{
addGrInformation(gr.f, assmblyName)
}
\arguments{
\item{gr.f}{A genomic range.}

\item{assmblyName}{human genome assembly name i.e., hg19 or hg38 or t2t.}
}
\value{
Input genomic ranges with added assembly information.
}
\description{
This function will add genomic length and assembly name to given genomic ranges (Human genome only). It will remove the non-standard chromosomes from genomic ranges and report the bad genomic ranges for the selected genome assembly.
}
\examples{

gr <- GRanges(seqnames = "chr1", strand = c("+", "-", "+"),  ranges = IRanges(start = c(1,110,105), end = c(100, 120, 150 )))
outGr=addGrInformation(gr.f=gr, assmblyName='hg19')



}
