% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_oligonucsCounts.R
\name{context_oligonucsCounts}
\alias{context_oligonucsCounts}
\title{oligo-nucleotide counts in within a genomic context.}
\usage{
context_oligonucsCounts(
  contextGr,
  oligoType,
  ignore.strand = FALSE,
  assmblyName = "hg19"
)
}
\arguments{
\item{contextGr}{GenomicRange object of the genomic context within which oligo-nucleotides has to be counted.}

\item{oligoType}{dinucs or trinucs or tetranucs}

\item{ignore.strand}{genomic context strand information should be considered. Default: FALSE}

\item{assmblyName}{human genome assembly name (hg19 or hg38 or t2t). Default: hg19}
}
\value{
a vector of oligo-nucleotide counts
}
\description{
oligo-nucleotide counts in within a genomic context.
}
\examples{
data(hg_38_gr)
oligonucs.Counts=context_oligonucsCounts(contextGr=hg_38_gr, oligoType='trinucs', ignore.strand=FALSE, assmblyName='hg38')



}
