% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDir_delIfExists.R
\name{createDir_delIfExists}
\alias{createDir_delIfExists}
\title{create a directory and delete if it already exists}
\usage{
createDir_delIfExists(dir)
}
\arguments{
\item{dir}{Name of the dir to be created}
}
\value{
dir created in the current diretory
}
\description{
create a directory and delete if it already exists
}
\examples{
createDir_delIfExists(dir='testDir')



}
