% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNASeqsForPattern.R
\name{DNASeqsForPattern}
\alias{DNASeqsForPattern}
\title{Generate all possible DNA sequences of a \href{https://genomevolution.org/wiki/index.php/Ambiguous_nucleotide}{Ambiguous nucleotide sequence}}
\usage{
DNASeqsForPattern(pat)
}
\arguments{
\item{pat}{Ambiguous nucleotide sequence}
}
\value{
A vector of all possible DNA sequences for a given Ambiguous nucleotide sequence
}
\description{
Generate all possible DNA sequences of a \href{https://genomevolution.org/wiki/index.php/Ambiguous_nucleotide}{Ambiguous nucleotide sequence}
}
\examples{
DNA_seqs=DNASeqsForPattern(pat='NYYN')
DNA_seqs

}
\seealso{
code{\link[DECIPHER]{Disambiguate}}
}
