% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyChrGranges.R
\name{emptyChrGranges}
\alias{emptyChrGranges}
\title{Create empty chromosome GenomicRange object for a given human genome assembly}
\usage{
emptyChrGranges(assmblyName)
}
\arguments{
\item{assmblyName}{hg19 or hg38 or t2t}
}
\value{
empty chromosome GenomicRange
}
\description{
Create empty chromosome GenomicRange object for a given human genome assembly (for standard chromosomes )
}
\examples{
hg_19_Chr.gr=emptyChrGranges('hg19')
hg_19_Chr.gr


}
