% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gbin_ByCGcnts.R
\name{Gbin_ByCGcnts}
\alias{Gbin_ByCGcnts}
\title{Creates GenomicRanges object of bins of user specified CpG counts for human genome.}
\usage{
Gbin_ByCGcnts(CGs_perBin, assmblyName, addSeq = FALSE)
}
\arguments{
\item{CGs_perBin}{CpG counts in a genomic block/bin. (even number)}

\item{assmblyName}{hg19 or hg38 or t2t}

\item{addSeq}{if sequence of the bin is required (Default: FALSE)}
}
\value{
GenomicRanges object of bins with user specified CpG counts
}
\description{
Creates a GenomicRange of genomic blocks of user specified CpG counts. (human genome)
}
\examples{
hg_19_CpGBins.gr=Gbin_ByCGcnts(CGs_perBin=1000, assmblyName='hg19' )
hg_19_CpGBins.gr


}
