% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGbins.R
\name{getGbins}
\alias{getGbins}
\title{Create GenomicRanges object of given the bin size  for human genome.}
\usage{
getGbins(assmblyName, binSize = 1000, chrName = chrName)
}
\arguments{
\item{assmblyName}{hg19 or hg38 or t2t}

\item{binSize}{size of the genomic block}

\item{chrName}{name of the chromosome ((UCSC format)), if not given then all chromosomes are considered (optional)}
}
\value{
GenomicRanges object of given the bin size
}
\description{
Create GenomicRanges object of given the bin size  for human genome.
}
\examples{
hg_19_Bins.gr=getGbins(assmblyName='hg19', binSize=1000 )
hg_19_Bins.gr


}
