% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPromGRange.R
\name{getPromGRange}
\alias{getPromGRange}
\title{Get genomic range of ProteinCodingGenes promoters (upst to -200)}
\usage{
getPromGRange(upst = 500, assmblyName)
}
\arguments{
\item{upst}{Upstream of the promoter. Promoter region will be "upst" to -200 to TSS. DEFAULT value is 500}

\item{assmblyName}{human genome assembly name i.e., hg19 or hg38}
}
\value{
genomic ranges for protein coding genes and their promoter for a give assembly.
}
\description{
This will return a promoters / genes genomic ranges for a given human assembly.
}
\examples{


outGr=getPromGRange(upst=500, assmblyName='hg19')
outGr$proteinCoding.proms
outGr$proteinCoding.genes



}
