% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeVariables.R
\name{largeVariables}
\alias{largeVariables}
\title{Memory usage of large variable in workspace}
\usage{
largeVariables(n = 3)
}
\arguments{
\item{n}{Number of  variables  needs to be listed ( decreasing memory storage order)}
}
\value{
DataFrame of top memory variables
}
\description{
Memory of variable in decreasing order
}
\examples{

largeVariables(n=5)




}
