% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftoverGR.R
\name{liftoverGR}
\alias{liftoverGR}
\title{Liftover genomic range  (Human genome)}
\usage{
liftoverGR(gr.f, liftFromTo)
}
\arguments{
\item{gr.f}{A genomic range.}
}
\value{
Liftovered GenomicRange object.
}
\description{
Liftover genomic range Human genome and will also add  assembly information to the output genomic ranges. Available for hg19, hg38, t2t genome assemblies. (filter out non-standard chromosomes)
}
\examples{
data(hg_38_gr)
outGr_h19=liftoverGR(gr.f=hg_38_gr, liftFromTo='hg38.To.hg19')
outGr_h19


}
