% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTracks_of_grangesList.R
\name{makeTracks_of_grangesList}
\alias{makeTracks_of_grangesList}
\title{make tracks list from GenomicRange lists.}
\usage{
makeTracks_of_grangesList(
  grlist,
  location,
  assmblyName = "hg19",
  if_plain = FALSE
)
}
\arguments{
\item{grlist}{list of genomic range objects.}

\item{location}{a list of genomic location with fields ´chr´, ´from´ and ´to´. Optional: Must be given if you choose ´if_plain=FALSE´}

\item{assmblyName}{human genome assembly name (hg19 or hg38 or t2t). Default: hg19}

\item{if_plain}{: ´TRUE´ then it will return only tracks of given genomic range. Default: FALSE: Tracks of the given chromosomes will also be returned.}
}
\value{
list of tracks for input list of genomic ranges objects.
}
\description{
Make tracks from genomic ranges objects list.
}
\examples{


data(cpgIslands)
mygrlist=list(gr1=cpgIslands,gr2=cpgIslands)
loc=list(chr='chr7',from=26700000, to=26750000)
tracks=makeTracks_of_grangesList(grlist=mygrlist, location=loc, assmblyName='hg19')
plotTracks(tracks, from = loc$from, to = loc$to)

# Example 2, only genomic ranges no annotation

gr1 <- GRanges(seqnames = "chr1", strand = c("+", "-", "+"),  ranges = IRanges(start = c(1,110,105), end = c(100, 120, 150 )))
gr2 <- GRanges(seqnames = "chr1", strand = c("+", "-", "+", "+"),  ranges = IRanges(start = c(1,12,60, 105), end = c(25, 50, 70, 115 )))
plotTracks(makeTracks_of_grangesList(list(gr1=gr1,gr2=gr2), if_plain=TRUE), shape="box")

}
