% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctOverlap_Of_FirstGrToSecondGr.R
\name{pctOverlap_Of_FirstGrToSecondGr}
\alias{pctOverlap_Of_FirstGrToSecondGr}
\title{Percent overlap of a GenomicRange with other}
\usage{
pctOverlap_Of_FirstGrToSecondGr(FirstContext, SecondContext)
}
\arguments{
\item{FirstContext}{GenomicRange object. (query: overlapped)}

\item{SecondContext}{GenomicRange object. (overlapped to)}
}
\value{
Percent overlap of first genomic range to second genomic range
}
\description{
Percent overlap of a genomic range with other
}
\examples{
gr1 <- GRanges(seqnames = "chr1", strand = c("+", "-", "+"),  ranges = IRanges(start = c(1,110,105), end = c(100, 120, 150 )))
gr2 <- GRanges(seqnames = "chr1", strand = c("+", "-", "+", "+"),  ranges = IRanges(start = c(1,12,60, 105), end = c(25, 50, 70, 115 )))
pctOverlap_Of_FirstGrToSecondGr(FirstContext=gr1, SecondContext=gr2)



}
